Extract parts of local signer code into signer
There are certain helper functions that will be useful for a wide range
of potential signers. These components have been extracted from the
local signer and placed into the signer package.
